+++ crates/tests/databases-tests/src/postgres/query_tests.rs
@@ -70,6 +70,16 @@ mod basic {
     }
 
     #[tokio::test]
+    async fn select_discoverable_composite_column() {
+        let result = run_query(
+            create_router().await,
+            "select_discoverable_composite_column",
+        )
+        .await;
+        insta::assert_json_snapshot!(result);
+    }
+
+    #[tokio::test]
     async fn select_composite_column_complex() {
         let result = run_query(create_router().await, "select_composite_column_complex").await;
         insta::assert_json_snapshot!(result);
@@ -96,6 +106,12 @@ mod basic {
     #[tokio::test]
     async fn select_nested_column_complex() {
         let result = run_query(create_router().await, "select_nested_column_complex").await;
+        insta::assert_json_snapshot!(result);
+    }
+
+    #[tokio::test]
+    async fn select_array_composite_field() {
+        let result = run_query(create_router().await, "select_array_composite_field").await;
         insta::assert_json_snapshot!(result);
     }
 }